Feature: mode fix

+++ StatisticalCalculatorOperations/Mode.cs
@@ -7,10 +7,10 @@ namespace StatisticalCalculatorOperations
   public class Mode
     {
   
-         public static double Findmode(int[] x)
+         public static double Findmode(double[] x)
         {
-            int high = x.OrderByDescending(n => n).First();
-            int mode = x.GroupBy(i => i)  //Grouping same items
+            double high = x.OrderByDescending(n => n).First();
+            double mode = x.GroupBy(i => i)  //Grouping same items
             .OrderByDescending(g => g.Count()) //now getting frequency of a value
             .Select(g => g.Key) //selecting key of the group
             .FirstOrDefault();   //Finally, taking the most frequent value
